separate render & recompose
