+++ src/packages/@ncigdc/modern_components/ClinicalAnalysis/ClinicalVariableCard/index.js
@@ -1,4 +1,4 @@
 import CategoricalVariableCard from './CategoricalVariableCard';
-import ContinuousVariableCard from './ContinuousVariableCard';
+import ContinuousVariableCard from './ContinuousCard';
 
 export { CategoricalVariableCard, ContinuousVariableCard };
